server/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
confusable-homoglyphs==3.2.0
cssselect==1.2.0
decorator==5.1.1
Django==4.1.5
django-cors-headers==3.13.0
django-registration==3.3
executing==1.2.0
idna==3.4
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
lxml==4.9.2
matplotlib-inline==0.1.6
numpy==1.24.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
readability-lxml==0.8.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
traitlets==5.8.0
urllib3==1.26.13
wcwidth==0.2.5
